% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funct-cmsy.R
\name{funct_cmsy}
\alias{funct_cmsy}
\title{run functional cmsy}
\usage{
funct_cmsy(
  catches = NA,
  catch_years = NA,
  start.yr = catch_years[1],
  end.yr = catch_years[length(catch_years)],
  stock = NA,
  region = NA,
  subregion = NA,
  common_name = NA,
  scientific_name = NA,
  bio_index = NA,
  r.low = 0.2,
  r.hi = 0.7,
  res = "medium",
  stb.low = NA,
  stb.hi = NA,
  int.yr = NA,
  intb.low = NA,
  intb.hi = NA,
  endb.low = NA,
  endb.hi = NA,
  q.start = NA,
  q.end = NA,
  e.creep = NA,
  btype = "None",
  force.cmsy = NA,
  comment = NA,
  source = NA,
  dataUncert = 0.3,
  sigmaR = 0.1,
  cor.log.rk = -0.607,
  rk.cor.beta = c(2.52, 3.37),
  nbk = 3,
  q.biomass.pr = c(0.9, 1.1),
  n = 20000,
  ni = 3,
  nab = 3,
  bw = 3,
  mgraphs = FALSE,
  e.creep.line = T,
  kobe.plot = FALSE,
  BSMfits.plot = FALSE,
  pp.plot = FALSE,
  retros = F,
  save.plots = F,
  close.plots = TRUE,
  select.yr = NA,
  lwr_k_buffer = 2,
  lwr_k_buffer2 = 2,
  upr_k_buffer2 = 10,
  upr_k_buffer = 6,
  cores = 1,
  plot_progress = FALSE,
  pt = FALSE,
  base_year = 1950,
  y2k_year = 2020
)
}
\arguments{
\item{catches}{a continuous vector of catches}

\item{stock}{the name of the stock}

\item{region}{the region the stock is in}

\item{subregion}{the subregion the stock is in}

\item{common_name}{the common name of the species}

\item{bio_index}{a vector of an abundance index}

\item{r.low}{lower bounds of intrinsic growth rate r}

\item{r.hi}{upper bound of intrinsic growth rate r}

\item{res}{resilience category for the stock}

\item{stb.low}{lower end b status in starting year}

\item{stb.hi}{upper end b status in starting year}

\item{int.yr}{year for intermediate estimate of stock status}

\item{intb.low}{lower end b status in intermediate year}

\item{intb.hi}{upper end b status in intermediate year}

\item{endb.low}{lower end b status in last year}

\item{endb.hi}{upper end b status in last year}

\item{q.start}{initial catchability}

\item{q.end}{final catchability}

\item{e.creep}{something to do with effort creep}

\item{btype}{the type of biomass index, one of None, CPUE, biomass}

\item{force.cmsy}{force use of cmsy instead of BSM, i think?}

\item{comment}{any comment that goes with the stock}

\item{source}{the source of any information, i think}

\item{dataUncert}{set observation error as uncertainty in catch - default is SD=0.3}

\item{sigmaR}{overall process error for CMSY; SD=0.1 is the default}

\item{cor.log.rk}{empirical value of log r-k correlation in 140 stocks analyzed with BSM (without r-k correlation)}

\item{rk.cor.beta}{beta.prior for rk cor+1}

\item{nbk}{Number of B/k priors to be used by BSM, with options 1 (first year), 2 (first & intermediate), 3 (first, intermediate & final bk priors)}

\item{q.biomass.pr}{if btype=="biomass" this is the prior range for q}

\item{n}{initial number of r-k pairs}

\item{ni}{iterations for r-k-startbiomass combinations, to test different variability patterns; no improvement seen above 3}

\item{nab}{recommended=5; minimum number of years with abundance data to run BSM}

\item{bw}{default bandwidth to be used by ksmooth() for catch data}

\item{mgraphs}{set to TRUE to produce additional graphs for management}

\item{e.creep.line}{set to TRUE to display uncorrected CPUE in biomass graph}

\item{kobe.plot}{set to TRUE to produce additional kobe status plot; management graph needs to be TRUE for Kobe to work}

\item{BSMfits.plot}{et to TRUE to plot fit diagnostics for BSM}

\item{pp.plot}{set to TRUE to plot Posterior and Prior distributions}

\item{retros}{leave as TRUE for retrospective bias plot}

\item{select.yr}{option to display F, B, F/Fmsy and B/Bmsy for a certain year; default NA}

\item{years}{a continuous vector of years corresponding to catches}

\item{ScientificName}{the scientific name of the species}
}
\value{
a dataframe of estimates by year
}
\description{
ported from https://github.com/SISTA16/cmsy on August 5 2020
No changes to any methods made
}
